+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to this extension pack will be documented in this file.
 
 ## 0.0.1
 
+- Added Table runner parser foundation [#1](https://github.com/rodrigotol/vscode-go-pack-go/pull/1)
 - Initial project scaffold.